+++ view/adminhtml/requirejs-config.js
@@ -4,7 +4,8 @@ var config = {
         'Vue': 'Snowdog_Menu/js/lib/vue',
         'vue': 'Snowdog_Menu/js/lib/require-vuejs',
         'Vddl': 'Snowdog_Menu/js/lib/vddl',
-        'vue-select': 'Snowdog_Menu/js/lib/vue-select'
+        'vue-select': 'Snowdog_Menu/js/lib/vue-select',
+        'vue-treeselect': 'Snowdog_Menu/js/lib/vue-treeselect'
     },
     shim: {
         'Vue': { 'exports': 'Vue' }